still reducing differences against official lwip same tests working (STA, telnet, echo) Signed-off-by: David Gauchard <gauchard@laas.fr>
